+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Bugfixes
 
 * Fix reading in NONMEM models with TIME column having hh:mm format (but no DATx column present)
 * Fix NONMEM parsing issues where having WRES in $TABLE could lead to parsing other columns incorrectly
+* Let translate_nmtran_time return the input model if the input model has no dataset
 
 1.2.0 (2024-08-22)
 ------------------